More Fortnite classes Fixed ByteProperty for tagged serialization

+++ src/main/kotlin/me/fungames/jfortniteparse/ue4/assets/JsonSerializer.kt
@@ -79,7 +79,7 @@ object JsonSerializer {
             is Long -> JsonPrimitive(ob)
 
             //Structs
-            is FIntPoint -> jsonObject("X" to ob.x.toLong(), "Y" to ob.y.toLong())
+            is FIntPoint -> jsonObject("X" to ob.x, "Y" to ob.y)
             is FGameplayTagContainer -> ob.gameplayTags.toJson(context)
             is FColor -> jsonObject("R" to ob.r.toShort(), "B" to ob.b.toShort(), "G" to ob.g.toShort(), "A" to ob.a)
             is FLinearColor -> jsonObject("R" to ob.r, "B" to ob.b, "G" to ob.g, "A" to ob.a)
